Update model
Updated the model to include Anthology, Version, and Mode as their own
tables rather than text fields for purposes of expanding compatibility
to support plugins from translationRecorder.

Character fields max_length has been set to 255 everywhere with the
exception of slugs which is 50. Default is 0 so something has to be set.

__unicode__() has been changed to __str__() to move to python3 from
python2

null and blank being set to true has been removed where nulls and blanks
should not be allowed, such as a project having a null or blank book.

Unique constraints fixed where the uniqueness should be based on
pairings of fields rather than just one.